% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allscr.R
\docType{data}
\name{allscr}
\alias{allscr}
\title{ASCI scores for all sites in California}
\format{A \code{\link{asci}} object}
\usage{
allscr
}
\description{
ASCI scores for all sites in California
}
\examples{
data(allscr)
}
\keyword{datasets}
