% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCI.R
\name{ASCI}
\alias{ASCI}
\title{Score samples using the ASCI tool}
\usage{
ASCI(taxain, sitein, tax = c("diatoms", "sba", "hybrid"), ...)
}
\arguments{
\item{taxain}{\code{data.frame} for input taxonomy data}

\item{sitein}{\code{data.frame} for input site data}

\item{tax}{chr string indicating output to return from a specific taxa, must one to many of \code{'diatoms'}, \code{'sba'}, or \code{'hybrid'}, defaults to all}

\item{...}{additional arguments passed to other funcions, e.g., \code{\link{rfpred}}}
}
\value{
A \code{\link{asci}} object with specific methods.  See the examples for accessing.
}
\description{
Score samples using the ASCI tool
}
\details{
Three taxonomy types are scored for the pMMI: diatoms, soft-bodied algae, and hybrid. This function combines output from the \code{\link{pmmifun}} function in a user-friendly format.
}
\examples{
results <- ASCI(demo_algae_tax, demo_algae_sitedata)
scores(results)
Supp1_mmi(results)
}
\seealso{
\code{\link{pmmifun}}
}
