% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcgis.R
\name{calcgis}
\alias{calcgis}
\title{Calculate missing station data}
\usage{
calcgis(station)
}
\arguments{
\item{station}{\code{data.frame} for input station data}
}
\value{
The original station data with calculated fields where applicable.
}
\description{
Calculate missing station data for \code{XerMtn} and \code{CondQR50}
}
\details{
\code{XerMtn} is calculated from \code{PSA6C} if the latter is present with no \code{NA} value. If \code{XerMtn} is 
present with \code{NA} values, \code{PSA6C} is used if present with no \code{NA} values. \code{XerMtn = 1} if
\code{PSA6C} is \code{'SN', 'NC'}, otherwise \code{XerMtn = 0} if \code{PSA6C} is \code{'CH', 'CV', 'DM', 'SC'}.

\code{CondQR50} is calculated from the \code{\link[quantregForest]{quantregForest}} object in \code{\link{rfmods}}
if all predictors are present in stations and no missing values are in the predictors.  Similar to \code{XerMtn}, 
\code{NA} values will be predicted per row only if the column already exists.  The required predictors are \code{'CaO_Mean', 
'MgO_Mean', 'S_Mean', 'UCS_Mean', 'LPREM_mean', 'AtmCa', 'AtmMg', 'AtmSO4', 'MINP_WS', 'MEANP_WS', 'SumAve_P',
'TMAX_WS', 'XWD_WS', 'MAXWD_WS', 'LST32AVE', 'BDH_AVE', 'KFCT_AVE', 'PRMH_AVE'}.
}
\examples{

# this calculates CondQR50 and XerMtn
calcgis(demo_station)

\dontrun{
# get XerMtn from PSA6c
tmp <- demo_station[, !names(demo_station) \%in\% 'XerMtn']
calcgis(tmp)

# error, cannot get XerMtn if PSA6C not found
tmp <- demo_station[, !names(demo_station) \%in\% c('XerMtn', 'PSA6C')]
calcgis(tmp)

# get conductivity
tmp <- demo_station
calcgis(tmp)

# get conductivity for only NA 
tmp <- demo_station
tmp$CondQR50[1] <- 200
calcgis(tmp)

# error, cannot calculate conductivity if missing predictors
tmp <- demo_station[, !names(demo_station) \%in\% c('TMAX_WS', 'AtmSO4')]
calcgis(tmp)

# error, cannot calculate conductivity if missing values in predictors
tmp <- demo_station
tmp$MINP_WS[2] <- NA
tmp$AtmSO4[3] <- NA
calcgis(tmp)
}
}
\seealso{
\code{\link{chkinp}}, \code{\link{getids}}
}
