% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkinp.R
\name{chkinp}
\alias{chkinp}
\title{Check input taxonomy and site data}
\usage{
chkinp(taxain, sitein)
}
\arguments{
\item{taxain}{\code{data.frame} for input taxonomy data}

\item{sitein}{\code{data.frame} for input site data}
}
\value{
\code{NULL} if all checks are met, otherwise an informative error message is returned
}
\description{
Check input taxonomy and site data for required information
}
\details{
The following is checked:
\itemize{
\item Required columns in taxonomy data: StationCode, SampleDate, Replicate,SampleTypeCode, BAResult, Result, FinalID
}
}
\examples{
chkinp(demo_algae_tax)
}
