% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_algae_sitedata.R
\docType{data}
\name{demo_algae_sitedata}
\alias{demo_algae_sitedata}
\title{Example data for environmental information at each site.}
\format{A \code{\link[base]{data.frame}} with 4 rows and 346 variables.}
\usage{
demo_algae_sitedata
}
\description{
An example dataset for environmental information at each site.  Sites match those in \code{\link{demo_algae_tax}}.
}
\examples{
data(demo_algae_sitedata)
}
\keyword{datasets}
