% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_station.R
\docType{data}
\name{demo_station}
\alias{demo_station}
\title{Example station}
\format{A \code{\link[base]{data.frame}} with 3 rows and 26 variables:}
\usage{
demo_station
}
\description{
An example dataset for station GIS data. Data are in wide format, one row per station with multiple fields per row.
}
\examples{
data(demo_station)
}
\keyword{datasets}
