% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diatom_rf_oe.R
\docType{data}
\name{diatom_rf_oe}
\alias{diatom_rf_oe}
\title{Diatom random forest model}
\format{A \code{\link[randomForest]{randomForest}} model for diatom communities.}
\usage{
diatom_rf_oe
}
\description{
Diatom random forest model used for prediction in \code{\link{oefun}}
}
\details{
The object is used internally in \code{\link{oefun}}.
}
\examples{
data(diatom_rf_oe)
}
\keyword{datasets}
