% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.R
\name{getids}
\alias{getids}
\title{Assign unique sample id}
\usage{
getids(datin, concatenate = T)
}
\arguments{
\item{datin}{\code{data.frame} of taxonomy or site data}

\item{concatenate}{logical indicating if \code{SampleID} is created or parsed}
}
\value{
The original input with a new \code{SampleID} column if \code{concatenate = TRUE}, or the original input with columns for \code{StationCode}, \code{Date}, and \code{Replicate} if \code{concatenate = FALSE}
}
\description{
Assing unique sample id to rows of taxonomy and site data
}
\details{
Assigns unique sample id based on a concatenation of StationCode, SampleDate, and Replicate, used internally in \code{\link{chkinp}} and \code{\link{ASCI}}
}
\examples{
getids(demo_algae_tax)

}
\seealso{
\code{\link{chkinp}}
}
