% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{hybrid}
\alias{hybrid}
\title{Calculate ASCI metrics for diatoms}
\usage{
hybrid(algae, gismetrics)
}
\arguments{
\item{algae}{\code{data.frame} input taxonomy data, with traits and indicators attached.}

\item{gismetrics}{\code{data.frame} the dataframe with the gis predictors}
}
\value{
a dataframe in wide format with the ASCI metrics for diatoms
}
\description{
Calculate ASCI metrics for diatoms
}
