% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_rf_oe.R
\docType{data}
\name{hybrid_rf_oe}
\alias{hybrid_rf_oe}
\title{Hybrid random forest model}
\format{A \code{\link[randomForest]{randomForest}} model for hybrid algal communities.}
\usage{
hybrid_rf_oe
}
\description{
Hybrid random forest model used for prediction in \code{\link{oefun}}
}
\details{
The object is used internally in \code{\link{oefun}}.
}
\examples{
data(hybrid_rf_oe)
}
\keyword{datasets}
