% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmi_calmetrics.R
\name{mmi_calcmetrics}
\alias{mmi_calcmetrics}
\title{Calculate ASCI metrics}
\usage{
mmi_calcmetrics(taxa = c("diatoms", "sba", "hybrid"), tax_dat, stations)
}
\arguments{
\item{taxa}{chr string indicating taxa to use to calculate metrics}

\item{tax_dat}{data.frame of taxonomy data for diatoms, soft-bodied algae, or hybrid}

\item{stations}{data.frame of site data}
}
\value{
a data.frame of scores for the relevant taxa
}
\description{
Calculate ASCI metrics
}
\details{
\code{tax_dat} is converted to \code{taxonomy_pa} for presence/absence
}
\examples{
\dontrun{
mmi_calcmetrics('diatoms', bugs.d.m)
}



}
