% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmilkup.R
\docType{data}
\name{mmilkup}
\alias{mmilkup}
\title{Lookup tables for MMI calculations}
\format{
A \code{\link[base]{list}} with six elements named \code{traits}, \code{d.win}, \code{sba.win}, \code{hybrid.win}, \code{omni.ref}, and \code{indicators}.
}
\usage{
mmilkup
}
\description{
Lookup tables for MMI calculations
}
\details{
Each element is a \code{\link[base]{data.frame}}.
}
\examples{
data(mmilkup)
}
\keyword{datasets}
