% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oefun.R
\name{oefun}
\alias{oefun}
\title{Run the ASCI O/E}
\usage{
oefun(taxain, sitein)
}
\arguments{
\item{taxain}{chr string for path to input taxonomy data}

\item{sitein}{chr string for path to input site data}
}
\value{
A list with three element named as \code{Scores.diatoms}, \code{Scores.sba}, and \code{Scores.hybrid}
}
\description{
Run the ASCI O/E index for diatoms, soft-bodied algae, and hyrids
}
\details{
Three index scores are calculated and returned as a named list
}
\examples{

oefun('demoalgae_tax.csv', 'demo_algae_sitedata.csv')
}
