% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oelkup.R
\docType{data}
\name{oelkup}
\alias{oelkup}
\title{Lookup tables for O/E calculations}
\format{A \code{\link[base]{list}} with six elements named \code{diatoms.bugs.rc}, \code{diatoms.stations.rc}, \code{sba.bugs.rc}, \code{sba.stations.rc}, \code{hybrid.bugs.rc}, and \code{hybrid.stations.rc}.}
\usage{
oelkup
}
\description{
Lookup tables for O/E calculations used in \code{\link{oefun}}
}
\details{
Each element is a \code{\link[base]{data.frame}} of taxonomic ('bugs') or site ('stations') information for diatoms, soft-bodied algae, or hyrid taxa at reference calibration sites.
}
\examples{
data(oelkup)
}
\seealso{
\code{\link{oefun}}
}
\keyword{datasets}
