% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asci-methods.R
\docType{methods}
\name{perf}
\alias{perf}
\alias{perf,asci-method}
\title{Get performance measures for ASCI}
\usage{
perf(object)

\S4method{perf}{asci}(object)
}
\arguments{
\item{object}{\code{\link{asci}} object for statewide results}
}
\value{
A \code{data.frame} of performance measures by index, index group (diatoms, soft-bodied algae, hybrid), predicted/null, and reference calibration/validation
}
\description{
Get performance measures for ASCI using statewide results
}
\details{
This functions estimates seven performance measures that evaluate index accuracy, precision, and responsiveness for calibration and validation datasets. The following are estimated:
\itemize{
\item{\code{ave}} {Average scores at reference sites}
\item{\code{fst}} {F statistic of reference scores between ecoregions}
\item{\code{vrs_nat}} {Variance of index scores explained by natural gradients at reference sites}
\item{\code{prc_amg}} {Precision among sites as standard deviation of scores at reference sites}
\item{\code{prc_wth}} {Precision within sites as standard deviation of within-site residuals for repeat site visits at reference sites}
\item{\code{res_tst}} {t-statistic as a measure of precision between scores at reference and stressed sites}
\item{\code{res_var}} {Variance of index scores explained by human-activity gradients at all sites as a measure of responsiveness}
}
This function requires site classifications as reference (calibration and validation), stressed, and repeat visits.  Regions for each site are also required.
}
\examples{
perf(allscr)
}
\seealso{
\code{\link{sitcat}}
}
