% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmi_calcmetrics.R
\name{pmmi_calcmetrics}
\alias{pmmi_calcmetrics}
\title{Calculate ASCI metrics}
\usage{
pmmi_calcmetrics(taxa = c("diatoms", "sba", "hybrid"))
}
\arguments{
\item{taxa}{chr string indicating taxa to use to calculate metrics}
}
\value{
a data.frame of scores for the relevant taxa
}
\description{
Calculate ASCI metrics
}
\examples{
pmmi_calcmetrics('diatoms')
}
