% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmilkup.R
\docType{data}
\name{pmmilkup}
\alias{pmmilkup}
\title{Lookup tables for pMMI calculations}
\format{A \code{\link[base]{list}} with six elements named \code{traits}, \code{d.win}, \code{sba.win}, \code{hybrid.win}, \code{quants}, and \code{indicators}.}
\usage{
pmmilkup
}
\description{
Lookup tables for pMMI calculations used in \code{\link{pmmifun}} and \code{\link{pmmi_calcmetrics}}
}
\details{
Each element is a \code{\link[base]{data.frame}} where all are used in \code{\link{pmmifun}}, except \code{indicators} which is used in \code{\link{pmmi_calcmetrics}}.
}
\examples{
data(pmmilkup)
}
\seealso{
\code{\link{pmmifun}}, \code{\link{pmmi_calcmetrics}}
}
\keyword{datasets}
