% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarify.R
\name{rarify}
\alias{rarify}
\title{Rarify a macroinvertebrate sample}
\usage{
rarify(inbug, sample.ID, abund, subsiz)
}
\arguments{
\item{inbug}{\code{data.frame} for input data}

\item{sample.ID}{chr string of column name with the sample ID}

\item{abund}{chr string of column name with the abundance data}

\item{subsiz}{numeric value indicating size of the subsample}
}
\description{
Rarify (subsample) a macroinvertebrate sample down to a fixed count
}
\note{
Version 1.0 06/10/05
}
\author{
John Van Sickle, USEPA, \mail{VanSickle.John@epa.gov}
}
