% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_out_top.R
\docType{data}
\name{rf_out_top}
\alias{rf_out_top}
\title{Random forest model for predicting diatom metrics}
\format{A \code{\link[randomForest]{randomForest}} model for diatom communities.}
\usage{
rf_out_top
}
\description{
Random forest model for predicting diatom metrics in \code{\link{pmmi_calcmetrics}}
}
\details{
The object is used internally in \code{\link{pmmi_calcmetrics}}.
}
\examples{
data(rf_out_top)
}
\keyword{datasets}
