% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfpred.r
\name{rfpred}
\alias{rfpred}
\title{Predict O/E scores for new sites
 
Predict O/E scores for a new set of sites based on a Random Forest predictive model}
\usage{
rfpred(bugcal.pa, grps.final, preds.final, ranfor.mod, prednew, bugnew,
  Pc = 0.5, Cal.OOB = FALSE, addnull = FALSE, bcpred = FALSE)
}
\arguments{
\item{bugcal.pa}{\code{data.frame} of taxonomy at calibration reference sites}

\item{grps.final}{\code{data.frame} of environmetnal data at calibration reference sites}

\item{preds.final}{chr string of predictors used in \code{ranfor.mod}}

\item{ranfor.mod}{\code{\link[randomForest]{randomForest}} model used for prediction}

\item{prednew}{\code{data.frame} containing predictor variables (columns) at all new sites/samples (rows) at which predictions are desired (e.g., 'test' sites)}

\item{bugnew}{\code{data.frame} of taxonomy presence/absence data at all new sites/samples (rows) at which predictions are desired (e.g., 'test' sites)}

\item{Pc}{numeric value for capture probability cutoff}

\item{Cal.OOB}{logical value default \code{FALSE}, set to \code{TRUE} only if you are predicting for the model calibration data and desire out-of-bag predictions}

\item{addnull}{logical indicating of null O/E scores are added to output}

\item{bcpred}{logical indicating if BC predictions are added to output}
}
\value{
function output is a list containing three elements
\itemize{ 
\item \code{OE.scores} A data frame for all samples, containing O, E, O/E, and O/E percentile from the predictive model (null model results can be included if \code{addnull = TRUE})
\item \code{Capture.Probs} Matrix of model-predicted capture (occurrence) probabilties for all taxa in all samples
\item \code{Group.Occurrnce.Probs} Matrix of predicted probabilities of occurrence for each sample in each calibration-site group
}
}
\description{
Predict O/E scores for new sites
 
Predict O/E scores for a new set of sites based on a Random Forest predictive model
}
\note{
Version 4.2 - Added option for out-of-bag predictions on CAL data. 12/30/10
}
\author{
J. Van Sickle, USEPA, \email{VanSickle.John@epa.gov}
}
