% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sba.R
\name{sba}
\alias{sba}
\title{Calculate ASCI metrics for soft body algae}
\usage{
sba(algae)
}
\arguments{
\item{algae}{\code{data.frame} input taxonomy data, with traits and indicators attached}
}
\value{
a dataframe in wide format with the ASCI metrics for diatoms
}
\description{
Calculate ASCI metrics for soft body algae
}
