% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sba_rf_oe.R
\docType{data}
\name{sba_rf_oe}
\alias{sba_rf_oe}
\title{Soft-bodied algae (sba) random forest model}
\format{A \code{\link[randomForest]{randomForest}} model for Soft-bodied algae communities.}
\usage{
sba_rf_oe
}
\description{
Soft-bodied algae (sba) random forest model used for prediction in \code{\link{oefun}}
}
\details{
The object is used internally in \code{\link{oefun}}.
}
\examples{
data(sba_rf_oe)
}
\keyword{datasets}
