% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Create scored metrics}
\usage{
score(metrics, assemblage)
}
\arguments{
\item{metrics}{mmi metrics dataframe, having metrics to be scored}

\item{assemblage}{chr string indicating the assemblage type - diatoms, sba or hybrid}
}
\value{
a data.frame scored
}
\description{
Create scored metrics
}
\examples{
\dontrun{
score(taxa = 'diatoms', bugs.d.m, d.results, omni.ref)
}
}
