% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcify.R
\name{smcify}
\alias{smcify}
\title{Re format ASCI output for SMC database}
\usage{
smcify(df.wide)
}
\arguments{
\item{df.wide}{The wide data that gets output from the ASCI function}
}
\value{
a data.frame long - the same ASCI data in long format to go to the SMC database tables
}
\description{
Re format ASCI output for SMC database
}
