% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE.R
\docType{data}
\name{STE}
\alias{STE}
\title{Taxonomic nomenclature}
\format{
A \code{\link[base]{data.frame}} of taxonomic lookup data used for verifying species names in sample data, 4961 rows and five columns.
}
\usage{
STE
}
\description{
Taxonomic nomenclature for comparison and standardization of sample data
}
\details{
The object is used internally in \code{\link{ASCI}}
}
\examples{
data(STE)
}
\keyword{datasets}
