% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-asci.R, R/asci-methods.R
\docType{class}
\name{asci-class}
\alias{asci-class}
\alias{asci}
\alias{show,asci-method}
\alias{scores}
\alias{scores,asci-method}
\alias{Supp1_mmi}
\alias{Supp1_mmi,asci-method}
\title{An S4 class to represent an \code{asci} object}
\usage{
\S4method{show}{asci}(object)

scores(object)

\S4method{scores}{asci}(object)

Supp1_mmi(object)

\S4method{Supp1_mmi}{asci}(object)
}
\arguments{
\item{object}{\code{asci} object created with \code{\link{ASCI}}}
}
\description{
The S4 \code{asci} object includes several methods shown below.
}
\section{Slots}{

\describe{
\item{\code{scores}}{\code{data.frame} of ASCI scores}

\item{\code{Supp1_mmi}}{\code{data.frame} of supplemental MMI metric scores}

\item{\code{taxa}}{character string of taxa returned with initial function call}
}}

\examples{
results <- ASCI(demo_algae_tax, demo_algae_sitedata)
scores(results)
Supp1_mmi(results)

}
