% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCI.R
\name{ASCI}
\alias{ASCI}
\title{Score samples using the ASCI tool}
\usage{
ASCI(taxain, sitein, tax = c("diatoms", "sba", "hybrid"), ...)
}
\arguments{
\item{taxain}{\code{data.frame} for input taxonomy data}

\item{sitein}{\code{data.frame} for input site data}

\item{tax}{chr string indicating output to return from a specific taxa, must one to many of \code{'diatoms'}, \code{'sba'}, or \code{'hybrid'}, defaults to all}

\item{...}{additional arguments passed to other funcions, e.g., \code{\link{rfpred}}}
}
\value{
A \code{asci} object as a \code{data.frame} with site scores for MMI and O/E indices.  Supplementary data are available as attributes named \code{Supp1_mmi}, \code{Supp1_OE}, and \code{Supp2_OE}.  See the examples for accessing.
}
\description{
Score samples using the ASCI tool
}
\details{
Two indices for three taxonomy types are scored, pMMI and O/E for diatoms, soft-bodied algae, and hybrid. This function combines output from the \code{\link{pmmifun}} and \code{\link{oefun}} functions in a user-friendly format.
}
\examples{
results <- ASCI(demo_algae_tax, demo_algae_sitedata)
attr(results, 'Supp1_mmi')
}
\seealso{
\code{\link{pmmifun}}, \code{\link{oefun}}
}
