% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCI.R
\name{ASCI}
\alias{ASCI}
\title{Score samples using the ASCI tool}
\usage{
ASCI(taxa, station)
}
\arguments{
\item{taxa}{\code{data.frame} for input taxonomy data}

\item{station}{\code{data.frame} for input station data}
}
\value{
A dataframe with all metrics calculated for each provided taxa
}
\description{
Score samples using the ASCI tool
}
\details{
One index for three taxonomy types are scored, MMI for diatoms, soft-bodied algae, and hybrid. 
If only soft-bodied algae or diatoms are present, only the respective ASCI and metrics are returned.
This function outputs the results of \code{\link{mmifun}} functions in a user-friendly format.
}
\examples{
# calculate all
ASCI(demo_algae_tax, demo_station)

# works if either soft-bodied or diatoms are missing
# remove diatoms from station sample 909M24937
tmp <- subset(demo_algae_tax, !(StationCode == '909M24937' & SampleTypeCode == 'Integrated'))
ASCI(tmp, demo_station)

# works if either soft-bodied or diatoms are missing
# remove soft-bodied from station sample 801M16916
tmp <- subset(demo_algae_tax, !(StationCode == '801M16916' & SampleTypeCode != 'Integrated'))
ASCI(tmp, demo_station)
}
\seealso{
\code{\link{mmifun}}
}
