% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCI.R
\name{ASCI}
\alias{ASCI}
\title{Score samples using the ASCI tool}
\usage{
ASCI(taxain, tax = c("diatoms", "sba", "hybrid"), ...)
}
\arguments{
\item{taxain}{\code{data.frame} for input taxonomy data}

\item{tax}{chr string indicating output to return from a specific taxa, must one to many of \code{'diatoms'}, \code{'sba'}, or \code{'hybrid'}, defaults to all}

\item{...}{additional arguments passed to other funcions}
}
\value{
A dataframe with all metrics calculated for each provided taxa
}
\description{
Score samples using the ASCI tool
}
\details{
One index for three taxonomy types are scored, MMI for diatoms, soft-bodied algae, and hybrid. 
This function outputs the reulsts of \code{\link{mmifun}} functions in a user-friendly format.
}
\examples{
results <- ASCI(demo_algae_tax)

}
\seealso{
\code{\link{mmifun}}
}
