% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCI.R
\name{ASCI}
\alias{ASCI}
\title{Score samples using the ASCI tool}
\usage{
ASCI(taxa, station, ...)
}
\arguments{
\item{taxa}{\code{data.frame} for input taxonomy data}

\item{station}{\code{data.frame} for input station data}

\item{...}{additional arguments passed to other functions}
}
\value{
A dataframe with all metrics calculated for each provided taxa
}
\description{
Score samples using the ASCI tool
}
\details{
One index for three taxonomy types are scored, MMI for diatoms, soft-bodied algae, and hybrid. 
This function outputs the results of \code{\link{mmifun}} functions in a user-friendly format.
}
\examples{
results <- ASCI(demo_algae_tax, demo_station)

}
\seealso{
\code{\link{mmifun}}
}
