% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkinp.R
\name{chkinp}
\alias{chkinp}
\title{Check input taxonomy data}
\usage{
chkinp(taxa, getval = FALSE)
}
\arguments{
\item{taxa}{\code{data.frame} for input taxonomy data}

\item{getval}{logical to return a vector of values not satisfied by checks, useful for data prep}
}
\value{
The original data with only relevant columns are returned if all checks are met, including a new column for \code{SampleID} (see \code{\link{getids}}).  An error message is returned if the datasets do not meet requirements or a vector of values that caused the error if \code{getval = TRUE}.  Site data will include only those sites in the taxonomic data.
}
\description{
Check input taxonomy for required information
}
\details{
The following are checked:
\itemize{
\item Required columns in taxonomy data: StationCode, SampleDate, Replicate, SampleTypeCode, BAResult, Result, FinalID
\item Taxonomic names are present in the \code{\link{STE}} reference file
\item Sites include both diatom and soft-bodied algae data (warning if not)
\item No missing abundance values for diatoms (for rarification)
}
}
\examples{
# all checks passed, data returned with SampleID
chkinp(demo_algae_tax)

# errors
\dontrun{
# missing columns in taxa data
tmp <- demo_algae_tax[, 1, drop = FALSE]
chkinp(tmp)
chkinp(tmp, getval = TRUE)

# incorrect taxonomy
tmp <- demo_algae_tax
tmp[1, 'FinalID'] <- 'asdf'
chkinp(tmp)
chkinp(tmp, getval = TRUE)

# missing diatom data at sites
tmp <- merge(demo_algae_tax, STE, all.x = T) \%>\%
  filter(Class \%in\% 'Bacillariophyceae')
chkinp(tmp)
chkinp(tmp, getval = TRUE)

# missing abundance data for diatoms
tmp <- demo_algae_tax
tmp$BAResult <- NA
chkinp(tmp)
chkinp(tmp, getval = TRUE)
}
}
\seealso{
\code{\link{getids}}
}
