% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_algae_tax.R
\docType{data}
\name{demo_algae_tax}
\alias{demo_algae_tax}
\title{Example data for algae taxonomy}
\format{
A \code{\link[base]{data.frame}} with 208 rows and 7 variables:
\describe{
  \item{\code{StationCode}}{chr}
  \item{\code{SampleDate}}{chr}
  \item{\code{Replicate}}{int}
  \item{\code{SampleTypeCode}}{chr}
  \item{\code{BAResult}}{num}
  \item{\code{Result}}{num}
  \item{\code{FinalID}}{chr}
}
}
\usage{
demo_algae_tax
}
\description{
An example dataset for algae taxonomy. Data are in long format, one row per sample
}
\examples{
data(demo_algae_tax)
}
\keyword{datasets}
