% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.R
\name{getids}
\alias{getids}
\title{Assign unique sample id}
\usage{
getids(datin)
}
\arguments{
\item{datin}{\code{data.frame} of taxonomy or site data}
}
\value{
The original input with a new 'SampleID' column
}
\description{
Assing unique sample id to rows of taxonomy and site data
}
\details{
Assigns unique sample id based on a concatenation of StationCode, SampleDate, and Replicate, used internally in \code{\link{chkinp}}
}
\examples{
getids(demo_algae_tax)

}
\seealso{
\code{\link{chkinp}}
}
