% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oefun.R
\name{oefun}
\alias{oefun}
\title{Run the ASCI O/E}
\usage{
oefun(taxain, sitein, ...)
}
\arguments{
\item{taxain}{\code{data.frame} for input taxonomy data}

\item{sitein}{\code{data.frame} for input site data}

\item{...}{additional arguments passed to \code{\link{rfpred}}}
}
\value{
A list with three elements named as \code{Scores.diatoms}, \code{Scores.sba}, and \code{Scores.hybrid}.  Each element includes an additional list of three data.frames names \code{OE.scores}, \code{Capture.Probs}, and \code{Group.Occurrence.Probs}.
}
\description{
Run the ASCI O/E index for diatoms, soft-bodied algae, and hybrids
}
\details{
Three index scores are calculated and returned as a named list
}
\examples{
taxain <- getids(demo_algae_tax)
sitein <- getids(demo_algae_sitedata)
oefun(taxain, sitein)
}
