% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oefun.R
\name{oefun}
\alias{oefun}
\title{Run the ASCI O/E}
\usage{
oefun(taxain, sitein)
}
\arguments{
\item{taxain}{chr string for path to input taxonomy data}

\item{sitein}{chr string for path to input site data}
}
\value{
A list with three elements named as \code{Scores.diatoms}, \code{Scores.sba}, and \code{Scores.hybrid}.  Each element includes an additional list of three data.frames names \code{OE.scores}, \code{Capture.Probs}, and \code{Group.Occurrence.Probs}.
}
\description{
Run the ASCI O/E index for diatoms, soft-bodied algae, and hybrids
}
\details{
Three index scores are calculated and returned as a named list
}
\examples{
oefun(demo_algae_tax, demo_algae_sitedata)
}
