% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmi_calcmetrics.R
\name{pmmi_calcmetrics}
\alias{pmmi_calcmetrics}
\title{Calculate ASCI metrics}
\usage{
pmmi_calcmetrics(taxa = c("diatoms", "sba", "hybrid"), tax_dat)
}
\arguments{
\item{taxa}{chr string indicating taxa to use to calculate metrics}

\item{tax_dat}{data.frame of taxonomy data for diatoms, soft-bodied algae, or hybrid}
}
\value{
a data.frame of scores for the relevant taxa
}
\description{
Calculate ASCI metrics
}
\details{
\code{tax_dat} is converted to \code{taxonomy_pa} for presence/absence
}
\examples{
\dontrun{
pmmi_calcmetrics('diatoms', bugs.d.m)
}
}
