% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmifun.R
\name{pmmifun}
\alias{pmmifun}
\title{Run the ASCI pMMI}
\usage{
pmmifun(taxain, sitein)
}
\arguments{
\item{taxain}{\code{data.frame} for input taxonomy data}

\item{sitein}{\code{data.frame} for input site data}
}
\value{
A list with three elements named as \code{d.results.scored}, \code{sba.results.scored}, and \code{hybrid.results.scored}. Each element is a \code{data.frame} with metric scores by site.
}
\description{
Run the ASCI pMMI index for diatoms, soft-bodied algae, and hybrids
}
\details{
Three index scores are calculated and returned as a named list
}
\examples{
taxain <- getids(demo_algae_tax)
sitein <- getids(demo_algae_sitedata)
pmmifun(taxain, sitein)
}
