% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmods.R
\docType{data}
\name{rfmods}
\alias{rfmods}
\title{Random forest models for metrics}
\format{A \code{\link[base]{list}} with six elements, one for each predictive metric and a final element for conductivity estimates:
\describe{
  \item{\code{diatoms.prop.spp.BCG12}}{randomForest}
  \item{\code{diatoms.prop.spp.Salinity.BF}}{randomForest}
  \item{\code{diatoms.prop.spp.Trophic.E}}{randomForest}
  \item{\code{hybrid.OxyRed.DO_30.richness}}{randomForest}
  \item{\code{hybrid.prop.spp.BCG4}}{randomForest}
  \item{\code{hybrid.Salinity.BF.richness}}{randomForest}
  \item{\code{cond.qrf}}{quantregForest}
}}
\usage{
rfmods
}
\description{
List of random forest models for predictive metrics
}
\examples{
data(rfmods)
}
\keyword{datasets}
