% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_metric.R
\name{score_metric}
\alias{score_metric}
\title{Title}
\usage{
score_metric(taxa, bugs.m, results.metric, omni.ref)
}
\arguments{
\item{taxa}{}

\item{bugs.m}{}

\item{results.metric}{}

\item{omni.ref}{}
}
\description{
Title
}
